requirements.txt
Flask>=2.2,<4.0
Pillow>=9.5,<11.0
python-dotenv>=1.0,<2.0

